Create a list of ranges objects from looping data

Uses tab delimited looping data in the form of BEDPE files to create custom GRanges loop objects

LoopsToRanges(..., loop_names = NULL, custom_cols = 0,
  custom_mcols = NULL)

Arguments

...
Any number of tab delimited loop data files in BEDPE format

loop_names
A character vector of names for the loop datasets (optional)

custom_cols
An integer indicating the number of extra columns in the BEDPE file (default = 0)

custom_mcols
An integer or vector of integers indicating which columns are used for metadata (optional)

Value

A LoopRanges class object: list of GRanges looping data objects

Examples

# Load loops into LoopRanges object 
ovary_loops <- system.file("extdata/loops", "ovary_hg19.bedpe",
package = "LoopRig", mustWork = TRUE)
spleen_loops <- system.file("extdata/loops", "spleen_hg19.bedpe",
package = "LoopRig", mustWork = TRUE)
LoopsToRanges(ovary_loops, spleen_loops, custom_cols = 0)

#> [[1]]
#> GRangesList object of length 2:
#> $Anchor 1 
#> GRanges object with 5648 ranges and 0 metadata columns:
#>          seqnames              ranges strand
#>             <Rle>           <IRanges>  <Rle>
#>      [1]     chr5   14140109-14150109      *
#>      [2]     chr5   29850107-29860107      *
#>      [3]     chr5 158137008-158147008      *
#>      [4]     chr5 158197008-158207008      *
#>      [5]     chr5 158167008-158177008      *
#>      ...      ...                 ...    ...
#>   [5644]     chr4   78621154-78631154      *
#>   [5645]     chr4   33601622-33611622      *
#>   [5646]     chr4   41452017-41462017      *
#>   [5647]     chr4   99321151-99331151      *
#>   [5648]     chr4 169501151-169511151      *
#> 
#> ...
#> <1 more element>
#> -------
#> seqinfo: 23 sequences from an unspecified genome; no seqlengths
#> 
#> [[2]]
#> GRangesList object of length 2:
#> $Anchor 1 
#> GRanges object with 13763 ranges and 0 metadata columns:
#>           seqnames              ranges strand
#>              <Rle>           <IRanges>  <Rle>
#>       [1]     chr5 142009565-142019565      *
#>       [2]     chr5 159437007-159447007      *
#>       [3]     chr5 159507007-159517007      *
#>       [4]     chr5 159517007-159527007      *
#>       [5]     chr5   95185704-95195704      *
#>       ...      ...                 ...    ...
#>   [13759]     chr4   38361621-38371621      *
#>   [13760]     chr4 103551157-103561157      *
#>   [13761]     chr4 103591157-103601157      *
#>   [13762]     chr4 140791154-140801154      *
#>   [13763]     chr4 103521157-103531157      *
#> 
#> ...
#> <1 more element>
#> -------
#> seqinfo: 23 sequences from an unspecified genome; no seqlengths
#> 
#> attr(,"class")
#> [1] "LoopRanges"